requirements.txt
bodhi-llm[all]
torch>=2.1.0
transformers>=4.50.0
pyyaml>=6.0
peft>=0.10.0
trl>=0.9.0,<0.12.0
datasets>=2.18.0
accelerate>=0.28.0
autoawq>=0.2.0
numpy
pandas
tqdm
